refactor<tel> moved some functions and changed in the styles
