+++ teleop/htm/static/JS/mobileController/mobileController_c_logic.js
@@ -11,7 +11,6 @@ function setMobileCommand(x, y) {
 	} else {
 		console.error('Invalid types for setMobileCommand:', { x, y });
 	}
-	console.log(x, y);
 }
 
 function addDataToMobileCommand(newData) {
