+++ teleop/htm/templates/userMenu/menu_logbox.html
@@ -1,4 +1,3 @@
-
 <div id="logbox_container">
 	<h1 class="title">Training Events</h1>
 
@@ -29,8 +28,9 @@
 					<th>Steer Intervention</th>
 					<th>Speed Intervention</th>
 					<th>Save</th>
-          <th>Latitude</th>
+					<th>Latitude</th>
 					<th>Longitude</th>
+					<!-- AP = AutoPilot -->
 					<th>AP Steer</th>
 					<th>AP Brake</th>
 					<th>AP Steer Confidence</th>
